% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity.R
\name{as.dissimilarity}
\alias{as.dissimilarity}
\alias{as_dissimilarity}
\alias{as.dissimilarity.matrix}
\alias{as.dissimilarity.dist}
\alias{as.dissimilarity.dissimilarity}
\title{Convert a dist or matrix object into a dissimilarity object}
\usage{
as.dissimilarity(x, ...)

as_dissimilarity(x, ...)

\method{as.dissimilarity}{matrix}(x, ...)

\method{as.dissimilarity}{dist}(x, ...)

\method{as.dissimilarity}{dissimilarity}(x, ...)
}
\arguments{
\item{x}{An object to coerce into a \code{dissimilarity} object.}

\item{...}{Further argument passed to the coercion method.}
}
\value{
A \code{dissimilarity} object.
}
\description{
Create a \code{dissimilarity} matrix from an existing distance matrix as \code{dist} object (e.g., from \code{\link[stats:dist]{stats::dist()}}, or \code{\link[vegan:vegdist]{vegan::vegdist()}}), or from a similarly shaped \code{matrix}object.
}
\examples{
SciViews::R
read("iris", package = "datasets") \%>.\%
  select(., -species) -> iris_num  # Only numeric columns from iris
# Construct a dist object
iris_dist <- dist(iris_num)
class(iris_dist)
# Convert it into a dissimilarity object
iris_dis <- as.dissimilarity(iris_dist)
class(iris_dis)
}
\seealso{
\code{\link[=dissimilarity]{dissimilarity()}}
}
