% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca.R
\name{ca}
\alias{ca}
\alias{autoplot.ca}
\alias{chart.ca}
\title{Correspondence Analysis (CA)}
\usage{
ca(obj, ...)

\method{autoplot}{ca}(
  object,
  choices = 1L:2L,
  type = c("screeplot", "altscreeplot", "biplot"),
  col = "black",
  fill = "gray",
  aspect.ratio = 1,
  repel = FALSE,
  ...
)

\method{chart}{ca}(
  data,
  choices = 1L:2L,
  ...,
  type = c("screeplot", "altscreeplot", "biplot"),
  env = parent.frame()
)
}
\arguments{
\item{obj}{A formula or a data frame with numeric columns, or a matrix, or a
table or xtabs two-way contingency table, see \code{\link[ca:ca]{ca::ca()}}. The formula version
allows to specify two categorical variables from a data frame as \code{~f1 + f2}.
The other versions analyze a two-way contingency table crossing two factors.}

\item{...}{Further arguments from \code{\link[ca:ca]{ca::ca()}} or for plot.}

\item{object}{A \strong{pcomp} object}

\item{choices}{Vector of two positive integers. The two axes to plot, by}

\item{type}{The type of plot to produce: \code{"screeplot"} or \code{"altscreeplot"}
for two versions of the screeplot, or \code{"biplot"} for the CA biplot.}

\item{col}{The color for the points representing the observations, black by
default.}

\item{fill}{The color to fill bars, gray by default}

\item{aspect.ratio}{height/width of the plot, 1 by default (for plots where
the ratio height / width does matter)}

\item{repel}{Logical. Should repel be used to rearrange points labels?
\code{FALSE}by default}

\item{data}{Idem}

\item{env}{The environment where to evaluate code, \code{parent.frame()} by
default, which should not be changed unless you really know what you are
doing!}
}
\value{
\code{\link[=pca]{pca()}} produces a \strong{ca} object.
}
\description{
\code{ca()} is a reexport of the function from the {ca} package, it
offers a \code{ca(formula, data, ...)} interface. It is supplemented here with
various \code{chart()} types.
}
\examples{
library(chart)
data(caith, package = "MASS")
caith # A two-way contingency table
class(caith) # in a data frame
caith_ca <- ca(caith)
summary(caith_ca)

chart$scree(caith_ca)
chart$altscree(caith_ca)

chart$biplot(caith_ca)
}
