% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle.R
\name{circle}
\alias{circle}
\title{Draw circles in a plot}
\usage{
circle(x = 0, y = 0, d = 1, col = 0, lwd = 1, lty = 1, ...)
}
\arguments{
\item{x}{The x coordinate of the center of the circle.}

\item{y}{The y coordinate of the center of the circle.}

\item{d}{The diameter of the circle.}

\item{col}{The color of the border of the circle.}

\item{lwd}{The width of the circle border.}

\item{lty}{The line type to use to draw the circle.}

\item{...}{More arguments passed to \code{\link[=symbols]{symbols()}}.}
}
\value{
This function returns \code{NULL}. It is invoked for it side effect of
adding a circle in a base R plot.
}
\description{
Add a circle in a base R plot, given its center (x and y
coordinates) and its diameter. In {exploreit}, this function can be used to
cut a circular dendrogram, see example.
}
\examples{
plot(x = 0:2, y = 0:2)
circle(x = 1, y = 1, d = 1, col = "red", lwd = 2, lty = 2)
}
\seealso{
\code{\link[=symbols]{symbols()}}
}
