% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\alias{cluster.default}
\alias{cluster.dist}
\alias{str.Cluster}
\alias{labels.Cluster}
\alias{nobs.Cluster}
\alias{predict.Cluster}
\alias{augment.Cluster}
\alias{plot.Cluster}
\alias{autoplot.Cluster}
\alias{chart.Cluster}
\title{Hierarchical Clustering Analysis}
\usage{
cluster(x, ...)

\method{cluster}{default}(x, ...)

\method{cluster}{dist}(x, method = "complete", fun = NULL, ...)

\method{str}{Cluster}(object, max.level = NA, digits.d = 3L, ...)

\method{labels}{Cluster}(object, ...)

\method{nobs}{Cluster}(object, ...)

\method{predict}{Cluster}(object, k = NULL, h = NULL, ...)

\method{augment}{Cluster}(x, data, k = NULL, h = NULL, ...)

\method{plot}{Cluster}(
  x,
  y,
  labels = TRUE,
  hang = -1,
  check = TRUE,
  type = "vertical",
  lab = "Height",
  ...
)

\method{autoplot}{Cluster}(
  object,
  labels = TRUE,
  type = "vertical",
  circ.text.size = 3,
  theme = theme_sciviews(),
  xlab = "",
  ylab = "Height",
  ...
)

\method{chart}{Cluster}(data, ..., type = NULL, env = parent.frame())
}
\arguments{
\item{x}{A \code{Dissimilarity} object.}

\item{...}{Further arguments for the methods (see their respective manpages).}

\item{method}{The agglomeration method used. \code{"complete"} by default. Other
options depend on the function \verb{fun =} used. For the default one, you can
also use \code{"single"}, \code{"average"}, \code{"mcquitty"}, \code{"ward.D"}, \code{"ward.D2"},
\code{"centroid"}, or \code{"median"}.}

\item{fun}{The function to use to do the calculation. By default, it is
\code{\link[fastcluster:hclust]{fastcluster::hclust()}}, an fast and memory-optimized version of the default
R function \code{\link[stats:hclust]{stats::hclust()}}. You can also use \code{\link[flashClust:flashClust]{flashClust::hclust()}},
\code{\link[cluster:agnes]{cluster::agnes()}}, \code{\link[cluster:diana]{cluster::diana()}}, as well as, any other function that
returns an \code{hclust}object, or something convertible to \code{hclust} with
\code{\link[=as.hclust]{as.hclust()}}. The default (\code{NULL}) means that the \strong{fastcluster}
implementation is used.}

\item{object}{A \code{cluster} object.}

\item{max.level}{The maximum level to present.}

\item{digits.d}{The number of digits to print.}

\item{k}{The number of clusters to get.}

\item{h}{The height where the dendrogram should be cut (give either \verb{k =} or
\verb{h = }, but not both at the same time).}

\item{data}{The original dataset}

\item{y}{Do not use it.}

\item{labels}{Should we show the labels (\code{TRUE} by default).}

\item{hang}{The fraction of the plot height at which labels should hang below
(by default, -1 meaning labels are all placed at the extreme of the plot).}

\item{check}{The validity of the \code{cluster} object is verified first to avoid
crashing R. You can put it at \code{FALSE} to speed up computation if you are
really sure your object is valid.}

\item{type}{The type of dendrogram, by default, \code{"vertical"}. It could also
be \code{"horizontal"} (more readable when there are many observations), or
\code{"circular"} (even more readable with many observations, but more difficult
to chose the cutting level).}

\item{lab}{The label of the y axis (vertical) or x axis (horizontal), by
default \code{"Height"}.}

\item{circ.text.size}{Size of the text for a circular dendrogram}

\item{theme}{The ggplot2 theme to use, by default, it is \code{\link[=theme_sciviews]{theme_sciviews()}}.}

\item{xlab}{Label of the x axis (nothing by default)}

\item{ylab}{Label of the y axis, by default \code{"Height"}.}

\item{env}{The environment where to evaluate formulas. If you don't
understand this, it means you should not touch it!}
}
\value{
A \code{Cluster} object inheriting from \code{hclust}. Specific methods are: \code{\link[=str]{str()}} (compact display of the object content), \code{\link[=labels]{labels()}} (get the labels for the observations), \code{\link[=nobs]{nobs()}} (number of observations), \code{\link[=predict]{predict()}} (get the clusters, given a cutting level), \code{\link[=augment]{augment()}} (add the groups to the original data frame or tibble), \code{\link[=plot]{plot()}} (create a dendrogram as base R plot), \code{\link[=autoplot]{autoplot()}} (create a dendrogram as a ggplot2), and \code{\link[=chart]{chart()}} (create a dendrogram as a chart variant of a ggplot2).
}
\description{
Hierarchical clustering is an agglomerative method that uses a
dissimilarity matrix to group individuals. It is represented by a dendrogram
that can be cut at a certain level to form the final clusters.
}
\examples{
SciViews::R
iris <- read("iris", package = "datasets")
iris_num <- iris[, -5] # Only numeric columns from iris
# Cluster the 150 flowers
iris_dis <- dissimilarity(iris_num, method = "euclidean", scale = TRUE)
(iris_clust <- cluster(iris_dis, method = "complete"))
str(iris_clust) # More useful
str(iris_clust, max.level = 3L) # Only the top of the dendrogram

# Dendrogram with base R graphics
plot(iris_clust)
plot(iris_clust, labels = FALSE, hang = 0.1)
abline(h = 3.5, col = "red")
# Horizontal dendrogram
plot(iris_clust, type = "horizontal", labels = FALSE)
abline(v = 3.5, col = "red")
# Circular dendrogram
plot(iris_clust, type = "circular", labels = FALSE)
circle(d = 3.5, col = "red")

# Chart version of the dendrogram
chart(iris_clust) +
  geom_dendroline(h = 3.5, color = "red")
# Horizontal dendrogram and without labels
chart$horizontal(iris_clust, labels = FALSE) +
  geom_dendroline(h = 3.5, color = "red")
# Circular dendrogram with labels
chart$circ(iris_clust, circ.text.size = 3) + # Abbreviate type and change size
  geom_dendroline(h = 3.5, color = "red")

# Get the clusters
predict(iris_clust, h = 3.5)
# Four clusters
predict(iris_clust, k = 4)
# Add the clusters to the data (.fitted column added)
augment(data = iris, iris_clust, k = 4)
}
\seealso{
\code{\link[=dissimilarity]{dissimilarity()}}, \code{\link[stats:hclust]{stats::hclust()}}, \code{\link[fastcluster:hclust]{fastcluster::hclust()}}
}
