% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity.R
\name{dissimilarity}
\alias{dissimilarity}
\alias{print.Dissimilarity}
\alias{labels.Dissimilarity}
\alias{nobs.Dissimilarity}
\alias{autoplot.Dissimilarity}
\alias{chart.Dissimilarity}
\title{Calculate a dissimilarity matrix}
\usage{
dissimilarity(
  data,
  formula = ~.,
  subset = NULL,
  method = "euclidean",
  scale = FALSE,
  rownames.col = "rowname",
  transpose = FALSE,
  fun = NULL,
  ...
)

\method{print}{Dissimilarity}(x, digits.d = 3L, rownames.lab = "labels", ...)

\method{labels}{Dissimilarity}(object, ...)

\method{nobs}{Dissimilarity}(object, ...)

\method{autoplot}{Dissimilarity}(
  object,
  order = TRUE,
  show.labels = TRUE,
  lab.size = NULL,
  gradient = list(low = "blue", mid = "white", high = "red"),
  ...
)

\method{chart}{Dissimilarity}(
  data,
  order = TRUE,
  show.labels = TRUE,
  lab.size = NULL,
  gradient = list(low = "blue", mid = "white", high = "red"),
  ...,
  type = NULL,
  env = parent.frame()
)
}
\arguments{
\item{data}{A data.frame, tibble or matrix.}

\item{formula}{A right-side only formula (\code{~ ...}) indicating which columns to keep in the data. The default one (\code{~ .}) keeps all columns.}

\item{subset}{An expression indicating which rows to keep from data.}

\item{method}{The distance (dissimilarity) method to use. By default, it is \code{"euclidean"}, but it can also be \code{"maximum"}, \code{"binary"}, \code{"minkowski"} from \code{\link[stats:dist]{stats::dist()}}, or \code{"bray"}, \code{"manhattan"}, \code{"canberra"}, \code{"clark"}, \code{"kulczynski"}, \code{"jaccard"}, \code{"gower"}, \code{"altGower"}, \code{"morisita"}, \code{"horm"}, \code{"mountfort"}, \code{"raup"}, \code{"binomial"}, \code{"chao"}, \code{"cao"}, \code{"mahalanobis"}, \code{"chisq"}, or \code{"chord"} from \code{\link[vegan:vegdist]{vegan::vegdist()}}, or any other distance from the function you provide in \verb{fun =}.}

\item{scale}{Do we scale (mean = 0, standard deviation = 1) the data before calculating the distance (\code{FALSE} by default)?}

\item{rownames.col}{In case the \code{data} object does not have row names (a \code{tibble} for instance), which column should be used for name of the rows?}

\item{transpose}{Do we transpose \code{data} first (to calculate distance between columns instead of rows)? By default, not (\code{FALSE}).}

\item{fun}{A function that does the calculation and return a \code{dist}-like object (similar to what \code{\link[stats:dist]{stats::dist()}}) provides. If \code{NULL} (by default), \code{\link[stats:dist]{stats::dist()}} or \code{\link[vegan:vegdist]{vegan::vegdist()}} is used, depending on \verb{method =}. Note that both functions calculate \code{"canberra"} differently, and in this case, it is the \code{\link[vegan:vegdist]{vegan::vegdist()}} version that is used by default. Other compatible functions: \code{\link[vegan:designdist]{vegan::designdist()}}, \code{\link[cluster:daisy]{cluster::daisy()}}, \code{\link[factoextra:dist]{factoextra::get_dist()}}, and probably more.}

\item{...}{Further parameters passed to the \verb{fun =} (see its man page) for \code{\link[=dissimilarity]{dissimilarity()}}, or further arguments passed to methods.}

\item{x, object}{A \code{Dissimilarity} object}

\item{digits.d}{Number of digits to print, by default, 3.}

\item{rownames.lab}{The name of the column containing the labels, by default \code{"labels"}.}

\item{order}{Do we reorder the lines and columns according to their resemblance (\code{TRUE} by default)?}

\item{show.labels}{Are the labels displayed on the axes (\code{TRUE} by default)?}

\item{lab.size}{Force the size of the labels (\code{NULL} by default for automatic size).}

\item{gradient}{The palette of color to use in the plot.}

\item{type}{The type of plot. For the moment, only one plot is possible and the default value (\code{NULL}) should not be changed.}

\item{env}{The environment where to evaluate the formula. If you don't understand this, you probably don't have to touch this arguments.}
}
\value{
An S3 object of class \code{c("Dissimilarity", "dist")}, thus inheriting from \code{dist}. A \code{Dissimilarity} object is better displayed (specific \code{print()} method), and has also dedicated methods \code{labels()} (get line and column labels), \code{nobs()} (get number of observations, that is, number of lines or columns), \code{autoplot()} (generate a ggplot2 from the matrix) and \code{chart()} (generate a chart version of the ggplot2).
}
\description{
Compute a distance matrix from all pairs of columns or rows in
a data frame, using a unified SciViews::R formula interface.
}
\examples{
SciViews::R
read("iris", package = "datasets") \%>.\%
  select(., -species) -> iris_num  # Only numeric columns from iris
# Compare the 150 flowers and nicely print the result
dissimilarity(iris_num, method = "manhattan")
# Compare the measurements by transposing and scaling them first
iris_dist <- dissimilarity(iris_num, method = "euclidean",
  scale = TRUE, transpose = TRUE)
iris_dist
class(iris_dist)
labels(iris_dist)
nobs(iris_dist)
# specific plots
autoplot(iris_dist)
chart(iris_dist, gradient = list(low = "green", mid = "white", high = "red"))
}
\seealso{
\code{\link[stats:dist]{stats::dist()}}, \code{\link[vegan:vegdist]{vegan::vegdist()}}, \code{\link[vegan:designdist]{vegan::designdist()}}, \code{\link[cluster:daisy]{cluster::daisy()}}, \code{\link[factoextra:dist]{factoextra::get_dist()}}
}
