% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_dendroline.R
\name{geom_dendroline}
\alias{geom_dendroline}
\title{Draw a line to cut a dendrogram}
\usage{
geom_dendroline(h, ...)
}
\arguments{
\item{h}{The height to cut the dendrogram.}

\item{...}{Further arguments passed to \code{\link[=geom_hline]{geom_hline()}} (this is really a convenience function that builds on it).}
}
\description{
Add a line (horizontal, vertical, or circular, depending on the
dendrogram type) at height \code{h} to depict where it is cut into groups.
}
\examples{
SciViews::R
iris <- read("iris", package = "datasets")
iris \%>.\%
  select(., -species) \%>.\%
  dissimilarity(.) \%>.\%
  cluster(.) ->
  iris_cluster
chart(iris_cluster) +
  geom_dendroline(h = 3, color = "red")
}
\seealso{
\code{\link[=geom_hline]{geom_hline()}}
}
