% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{mds}
\alias{mds}
\alias{plot.mds}
\alias{autoplot.mds}
\alias{chart.mds}
\alias{shepard}
\alias{plot.shepard}
\alias{autoplot.shepard}
\alias{chart.shepard}
\alias{augment.mds}
\alias{glance.mds}
\title{Multidimensional scaling or principal coordinates analysis}
\usage{
mds(
  dist,
  k = 2,
  type = c("metric", "nonmetric", "cmdscale", "wcmdscale", "sammon", "isoMDS",
    "monoMDS", "metaMDS"),
  p = 2,
  ...
)

\method{plot}{mds}(x, y, ...)

\method{autoplot}{mds}(object, labels, col, ...)

\method{chart}{mds}(data, labels, col, ..., type = NULL, env = parent.frame())

shepard(dist, mds, p = 2)

\method{plot}{shepard}(
  x,
  y,
  l.col = "red",
  l.lwd = 1,
  xlab = "Observed Dissimilarity",
  ylab = "Ordination Distance",
  ...
)

\method{autoplot}{shepard}(
  object,
  alpha = 0.5,
  l.col = "red",
  l.lwd = 1,
  xlab = "Observed Dissimilarity",
  ylab = "Ordination Distance",
  ...
)

\method{chart}{shepard}(
  data,
  alpha = 0.5,
  l.col = "red",
  l.lwd = 1,
  xlab = "Observed Dissimilarity",
  ylab = "Ordination Distance",
  ...,
  type = NULL,
  env = parent.frame()
)

\method{augment}{mds}(x, data, ...)

\method{glance}{mds}(x, ...)
}
\arguments{
\item{dist}{A \strong{dist} object from \code{\link[stats:dist]{stats::dist()}} or other compatible
functions like \code{\link[vegan:vegdist]{vegan::vegdist()}}, or a \strong{Dissimilarity} object, see
\code{\link[=dissimilarity]{dissimilarity()}}.}

\item{k}{The dimensions of the space for the representation, usually \code{k = 2}
(by default). It should be possible to use also \code{k = 3} with extra care and
custom plots.}

\item{type}{Not used}

\item{p}{For types \code{"nonmetric"}, \code{"metaMDS"}, \code{"isoMDS"}, \code{"monoMDS"} and
\code{"sammon"}, a Shepard plot is also precalculated. \code{p}is the power for Minkowski distance in the configuration scale. By default, \code{p = 2}. Leave it like that if you don't understand what it means see \code{\link[MASS:isoMDS]{MASS::Shepard()}}.}

\item{...}{More arguments (see respective \code{type}s or functions)}

\item{x}{Idem}

\item{y}{Not used}

\item{object}{An \strong{mds} object}

\item{labels}{Points labels on the plot (optional)}

\item{col}{Points color (optional)}

\item{data}{A data frame to augment with columns from the MDS analysis}

\item{env}{Not used}

\item{mds}{Idem}

\item{l.col}{Color of the line in the Shepard's plot (red by default)}

\item{l.lwd}{Width of the line in the Shepard"s plot (1 by default)}

\item{xlab}{Label for the X axis (a default value exists)}

\item{ylab}{Idem for the Y axis}

\item{alpha}{Alpha transparency for points (0.5 by default, meaning 50\%
transparency)}
}
\value{
A \strong{mds} object, which is a list containing all components from the
corresponding function, plus possibly \code{Shepard} if the Shepard plot is
precalculated.
}
\description{
Perform a PCoA (`type = "metric"). or other forms of MDS.
}
\examples{
library(chart)
data(iris, package = "datasets")
iris_num <- iris[, -5] # Only numeric columns
iris_dis <- dissimilarity(iris_num, method = "euclidean")

# Metric MDS
iris_mds <- mds$metric(iris_dis)
chart(iris_mds, labels = 1:nrow(iris), col = iris$Species)

# Non-metric MDS
iris_nmds <- mds$nonmetric(iris_dis)
chart(iris_nmds, labels = 1:nrow(iris), col = iris$Species)
glance(iris_nmds) # Good R^2
iris_sh <- shepard(iris_dis, iris_nmds)
chart(iris_sh) # Excellent matching + linear -> metric MDS is OK here
}
