% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{mfa}
\alias{mfa}
\alias{autoplot.MFA}
\alias{chart.MFA}
\title{Multiple Factor Analysis (MFA)}
\usage{
mfa(data, formula, nd = 5, suprow = NA, ..., graph = FALSE)

\method{autoplot}{MFA}(
  object,
  type = c("screeplot", "altscreeplot", "loadings", "scores", "groups", "axes",
    "contingency", "ellipses"),
  choices = 1L:2L,
  name = deparse(substitute(object)),
  col = "black",
  fill = "gray",
  title,
  ...,
  env
)

\method{chart}{MFA}(
  data,
  choices = 1L:2L,
  name = deparse(substitute(data)),
  ...,
  type = NULL,
  env = parent.frame()
)
}
\arguments{
\item{data}{A data frame}

\item{formula}{A formula that specifies the variables groups to consider
(see details)}

\item{nd}{Number of dimensions kept in the results (by default, 5)}

\item{suprow}{A vector indicating the row indices for the supplemental
individuals}

\item{...}{Additional arguments to \code{\link[FactoMineR:MFA]{FactoMineR::MFA()}} or to the plot}

\item{graph}{If \code{TRUE}a graph is displayed (\code{FALSE} by default)}

\item{object}{An \strong{MFA} object}

\item{type}{The type of plot to produce: \code{"screeplot"} or \code{"altscreeplot"}
for two versions of the screeplot, \code{"loadings"}, \code{"scores"}, \code{"groups"},
\code{"axes"}, \code{"contingency"} or \code{"ellipses"} for the different views of the
MFA.}

\item{choices}{Vector of two positive integers. The two axes to plot, by
default first and second axes.}

\item{name}{The name of the object (automatically defined by default)}

\item{col}{The color for the points representing the observations, black by
default.}

\item{fill}{The color to fill bars, gray by default}

\item{title}{The title of the plot (optional, a reasonable default is used)}

\item{env}{The environment where to evaluate code, \code{parent.frame()} by
default, which should not be changed unless you really know what you are
doing!}
}
\value{
An \strong{MFA} object
}
\description{
Analyze several groups of variables at once with supplementary
groups of variables or individuals. Each group can be numeric, factor or
contingency tables. Missing values are replaced by the column mean and
missing values for factors are treated as an additional level. This is a
formula interface to the \code{\link[FactoMineR:MFA]{FactoMineR::MFA()}} function.
}
\details{
The formula presents how the different columns of the data frame are grouped
and indicates the kind of sub-table they are and the name we give to them in
the analysis. So, a component of the formula for one group is
\code{n * kind \%as\% name} where \code{n} is the number of columns belonging to this
group, starting at column 1 for first group, \code{kind} is \code{std} for numeric
variables to be standardized and used as a PCA, \code{num} for numerical variables
to use as they are also as a PCA, \code{cnt} for counts in a contingency table to
be treated as a CA and \code{fct} for classical factors (categorical variables).
Finally, \code{name} is a (short) name you use to identify this group. The kind
may be omitted and it will be \code{std} by default. If \verb{\%as\% name} is omitted, a
generic name (group1, group2, group3, ...) is used. The complete formula is
the addition of the different groups to include in the analysis and the
subtraction of the supplementary groups not included in the analysis, like
\code{~n1*std \%as\% gr1 - n2*fct \%as\% gr2 + n3*num \%as\% gr3}, with groups "gr1" and
"gr3" included in the analysis and group "gr2" as supplemental. The total
\code{n1 + n2 + n3} must equal the number of columns in the data frame.
}
\note{
The symbols for the groups are different in \code{\link[=mfa]{mfa()}} and \code{\link[FactoMineR:MFA]{FactoMineR::MFA()}}).
To avoid further confusion, the symbols use three letters here:
\itemize{
\item \code{std} is the same as \code{s} in \code{MFA()}: "standardized" and is the default
\item \code{num} stands here for "numeric", thus continuous variables \code{c} in \code{MFA()}
\item \code{cnt} stands for "contingency" table and matches \code{f} in \code{MFA()}
\item \code{fct} stands for "factor", thus qualitative variables \code{n} in \code{MFA()}
}
}
\examples{
# Same example as in {FactoMineR}
library(chart)
data(wine, package = "FactoMineR")
wine_mfa <- mfa(data = wine,
  ~ -2*fct \%as\% orig +5 \%as\% olf + 3 \%as\% vis + 10 \%as\% olfag + 9 \%as\% gust - 2 \%as\% ens)
wine_mfa
summary(wine_mfa)

chart$scree(wine_mfa)
chart$altscree(wine_mfa)

chart$loadings(wine_mfa)
chart$scores(wine_mfa)
chart$groups(wine_mfa)

chart$axes(wine_mfa)
# No contingency group! chart$contingency(wine_mfa)
chart$ellipses(wine_mfa)
}
