% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\alias{autoplot.pcomp}
\alias{chart.pcomp}
\alias{augment.princomp}
\alias{tidy.princomp}
\alias{as.prcomp}
\alias{as.prcomp.default}
\alias{as.prcomp.prcomp}
\alias{as.prcomp.princomp}
\title{Principal Component Analysis (PCA)}
\usage{
pca(x, ...)

\method{autoplot}{pcomp}(
  object,
  type = c("screeplot", "altscreeplot", "loadings", "correlations", "scores", "biplot"),
  choices = 1L:2L,
  name = deparse(substitute(object)),
  ar.length = 0.1,
  circle.col = "gray",
  col = "black",
  fill = "gray",
  scale = 1,
  aspect.ratio = 1,
  repel = FALSE,
  labels,
  title,
  xlab,
  ylab,
  ...
)

\method{chart}{pcomp}(
  data,
  choices = 1L:2L,
  name = deparse(substitute(data)),
  ...,
  type = NULL,
  env = parent.frame()
)

\method{augment}{princomp}(x, data = NULL, newdata, ...)

\method{tidy}{princomp}(x, matrix = "u", ...)

as.prcomp(x, ...)

\method{as.prcomp}{default}(x, ...)

\method{as.prcomp}{prcomp}(x, ...)

\method{as.prcomp}{princomp}(x, ...)
}
\arguments{
\item{x}{A formula or a data frame with numeric columns, for \code{as.prcomp()},
an object to coerce into \strong{prcomp}.}

\item{...}{For \code{pca()}, further arguments passed to \code{\link[SciViews:pcomp]{SciViews::pcomp()}},
notably, \verb{data=}  associated with a formula, \verb{subset=}(optional),
\verb{na.action=}, \verb{method=} that can be \code{"svd"} or \code{"eigen"}. See
\code{\link[SciViews:pcomp]{SciViews::pcomp()}} for more details on these arguments.}

\item{object}{A \strong{pcomp} object}

\item{type}{The type of plot to produce: \code{"screeplot"} or \code{"altscreeplot"}
for two versions of the screeplot, \code{"loadings"}, \code{"correlations"}, or
\code{"scores"} for the different views of the PCA, or a combined \code{"biplot"}.}

\item{choices}{Vector of two positive integers. The two axes to plot, by
default first and second axes.}

\item{name}{The name of the object (automatically defined by default)}

\item{ar.length}{The length of the arrow head on the plot, 0.1 by default}

\item{circle.col}{The color of the circle on the plot, gray by default}

\item{col}{The color for the points representing the observations, black by
default.}

\item{fill}{The color to fill bars, gray by default}

\item{scale}{The scale to apply for annotations, 1 by default}

\item{aspect.ratio}{height/width of the plot, 1 by default (for plots where
the ratio height / width does matter)}

\item{repel}{Logical. Should repel be used to rearrange points labels?
\code{FALSE}by default}

\item{labels}{The label of the points (optional)}

\item{title}{The title of the plot (optional, a reasonable default is used)}

\item{xlab}{The label for the X axis. Automatically defined if not provided}

\item{ylab}{Idem for the Y axis}

\item{data}{The original data frame used for the PCA}

\item{env}{The environment where to evaluate code, \code{parent.frame()} by
default, which should not be changed unless you really know what you are
doing!}

\item{newdata}{A data frame with similar structure to \code{data} and new
observations}

\item{matrix}{Indicate which component should be be tidied. See
\code{\link[broom:tidy.prcomp]{broom::tidy.prcomp()}}}
}
\value{
\code{\link[=pca]{pca()}} produces a \strong{pcomp} object.
}
\description{
Principal Component Analysis (PCA)
}
\examples{
library(chart)
library(ggplot2)
data(iris, package = "datasets")
iris_num <- iris[, -5] # Only numeric columns
iris_pca <- pca(data = iris_num, ~ .)
summary(iris_pca)
chart$scree(iris_pca) # OK to keep 2 components
chart$altscree(iris_pca) # Different presentation

chart$loadings(iris_pca, choices = c(1L, 2L))
chart$scores(iris_pca, choices = c(1L, 2L), aspect.ratio = 3/5)
# or better:
chart$scores(iris_pca, choices = c(1L, 2L), labels = iris$Species,
  aspect.ratio = 3/5) +
  stat_ellipse()

# biplot
chart$biplot(iris_pca)
}
