% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreit.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{augment}
\alias{glance}
\title{Objects exported from other packages}
\seealso{
\code{\link[=augment.Cluster]{augment.Cluster()}}, \code{\link[=augment.kmeans]{augment.kmeans()}}

\code{\link[=glance.mds]{glance.mds()}}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{broom}{\code{\link[broom:reexports]{glance}}}

  \item{generics}{\code{\link[generics]{augment}}}
}}

