% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale}
\alias{scale}
\alias{scale.data.frame}
\alias{scale.tbl_df}
\alias{scale.data.table}
\title{Scale a data frame (data.frame, data.table or tibble's tbl_df)}
\usage{
\method{scale}{data.frame}(x, center = TRUE, scale = TRUE)

\method{scale}{tbl_df}(x, center = TRUE, scale = TRUE)

\method{scale}{data.table}(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{A data frame}

\item{center}{Are the columns centered (mean = 0)?}

\item{scale}{Are the column scaled (standard deviation = 1)?}
}
\value{
An object of the same class as \code{x}.
}
\description{
Center or scale all variables in a data frame. This takes a
data frame and return an object of the same class.
}
\examples{
data(trees, package = "datasets")
colMeans(trees)
trees2 <- scale(trees)
head(trees2)
class(trees2)
colMeans(trees2)
}
