% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreit-package.R
\docType{package}
\name{exploreit-package}
\alias{exploreit}
\alias{exploreit-package}
\title{exploreit: Exploratory Data Analysis in 'SciViews::R'}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Multivariate analysis and data exploration by using SciViews::R
  conventions.
}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

}
\keyword{internal}
