% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreit-package.R
\docType{package}
\name{exploreit-package}
\alias{exploreit}
\alias{exploreit-package}
\title{exploreit: 'SciViews' - Exploratory Data Analysis}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Multivariate analysis and data exploration by using 'SciViews::R' conventions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SciViews/exploreit}
  \item \url{https://www.sciviews.org/exploreit/}
  \item Report bugs at \url{https://github.com/SciViews/exploreit/issues}
}

}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

}
\keyword{internal}
